+++ src/form-validator/types.ts
@@ -3,4 +3,5 @@ export type FormErrors = Record<string, string[]>
 export interface FormRule {
   validate: (value: any, form: Record<string, any>) => boolean
   message: string
+  when?: (value: any, form: Record<string, any>) => boolean
 }